Optimize tests with ParameterizedTest, create kotlin files expected icons for tests
