+++ components/generator/imagevector/src/main/kotlin/io/github/composegears/valkyrie/generator/imagevector/spec/LazyDelegatePropertySpec.kt
@@ -12,36 +12,12 @@ import io.github.composegears.valkyrie.generator.ext.removeDeadCode
 import io.github.composegears.valkyrie.generator.ext.setIndent
 import io.github.composegears.valkyrie.generator.imagevector.ImageVectorSpecConfig
 import io.github.composegears.valkyrie.generator.imagevector.ImageVectorSpecOutput
-import io.github.composegears.valkyrie.generator.imagevector.util.iconPreviewSpec
-import io.github.composegears.valkyrie.generator.imagevector.util.iconPreviewSpecForNestedPack
-import io.github.composegears.valkyrie.generator.imagevector.util.imageVectorBuilderSpecs
 
-internal class LazyDelegatePropertySpec(
-    private val config: ImageVectorSpecConfig,
-) {
+internal class LazyDelegatePropertySpec(private val config: ImageVectorSpecConfig) {
 
     fun createAsLazyDelegateProperty(vector: Vector): ImageVectorSpecOutput {
-        val iconPackClassName = when {
-            config.iconPack.isEmpty() -> null
-            else -> {
-                if (config.iconNestedPack.isEmpty()) {
-                    ClassName(
-                        config.iconPackage,
-                        config.iconPack,
-                    )
-                } else {
-                    ClassName(
-                        config.iconPackage,
-                        config.iconPack,
-                    ).nestedClass(config.iconNestedPack)
-                }
-            }
-        }
-
-        val packageName = when {
-            config.iconNestedPack.isEmpty() -> config.iconPackage
-            else -> "${config.iconPackage}.${config.iconNestedPack.lowercase()}"
-        }
+        val iconPackClassName = config.resolveIconPackClassName()
+        val packageName = config.resolvePackageName()
 
         val fileSpec = fileSpecBuilder(
             packageName = packageName,
@@ -53,22 +29,11 @@ internal class LazyDelegatePropertySpec(
                     iconPackClassName = iconPackClassName,
                 ),
             )
-            apply {
-                if (config.generatePreview) {
-                    addFunction(
-                        funSpec = when {
-                            iconPackClassName != null -> iconPreviewSpecForNestedPack(
-                                iconPackClassName = iconPackClassName,
-                                iconName = config.iconName,
-                            )
-                            else -> iconPreviewSpec(
-                                iconPackage = packageName,
-                                iconName = config.iconName,
-                            )
-                        },
-                    )
-                }
-            }
+            addPreview(
+                config = config,
+                iconPackClassName = iconPackClassName,
+                packageName = packageName,
+            )
             setIndent()
         }
 
@@ -84,18 +49,7 @@ internal class LazyDelegatePropertySpec(
     ): PropertySpec = propertySpecBuilder(name = config.iconName, type = ClassNames.ImageVector) {
         receiver(iconPackClassName)
         val codeBlock = buildCodeBlock {
-            add(
-                imageVectorBuilderSpecs(
-                    iconName = when {
-                        config.iconNestedPack.isEmpty() -> config.iconName
-                        else -> "${config.iconNestedPack}.${config.iconName}"
-                    },
-                    vector = vector,
-                    path = {
-                        vector.nodes.forEach { node -> addVectorNode(node) }
-                    },
-                ),
-            )
+            addImageVectorBlock(config = config, vector = vector)
         }
 
         delegate(
