+++ components/generator/imagevector/src/test/kotlin/io/github/composegears/valkyrie/generator/imagevector/XmlToImageVectorTest.kt
@@ -1,0 +1,204 @@
+package io.github.composegears.valkyrie.generator.imagevector
+
+import assertk.assertThat
+import assertk.assertions.isEqualTo
+import io.github.composegears.valkyrie.generator.imagevector.common.ResourceUtils.getResourcePath
+import io.github.composegears.valkyrie.generator.imagevector.common.toResourceText
+import io.github.composegears.valkyrie.parser.IconParser
+import org.junit.jupiter.params.ParameterizedTest
+import org.junit.jupiter.params.provider.EnumSource
+
+class XmlToImageVectorTest {
+
+    @ParameterizedTest
+    @EnumSource(value = OutputFormat::class)
+    fun `generation without icon pack`(outputFormat: OutputFormat) {
+        val icon = getResourcePath("xml/ic_without_path.xml")
+        val parserOutput = IconParser.toVector(icon)
+        val output = ImageVectorGenerator.convert(
+            vector = parserOutput.vector,
+            kotlinName = parserOutput.kotlinName,
+            config = ImageVectorGeneratorConfig(
+                packageName = "io.github.composegears.valkyrie.icons",
+                packName = "",
+                nestedPackName = "",
+                outputFormat = outputFormat,
+                generatePreview = false,
+            ),
+        ).content
+
+        val expected = outputFormat.toResourceText(
+            pathToBackingProperty = "kt/backing/WithoutPath.kt",
+            pathToLazyProperty = "kt/lazy/WithoutPath.kt",
+        )
+        assertThat(output).isEqualTo(expected)
+    }
+
+    @ParameterizedTest
+    @EnumSource(value = OutputFormat::class)
+    fun `generation with nested icon pack`(outputFormat: OutputFormat) {
+        val icon = getResourcePath("xml/ic_without_path.xml")
+        val parserOutput = IconParser.toVector(icon)
+        val output = ImageVectorGenerator.convert(
+            vector = parserOutput.vector,
+            kotlinName = parserOutput.kotlinName,
+            config = ImageVectorGeneratorConfig(
+                packageName = "io.github.composegears.valkyrie.icons",
+                packName = "ValkyrieIcons",
+                nestedPackName = "Colored",
+                outputFormat = outputFormat,
+                generatePreview = false,
+            ),
+        ).content
+
+        val expected = outputFormat.toResourceText(
+            pathToBackingProperty = "kt/backing/Colored.WithoutPath.kt",
+            pathToLazyProperty = "kt/lazy/Colored.WithoutPath.kt",
+        )
+        assertThat(output).isEqualTo(expected)
+    }
+
+    @ParameterizedTest
+    @EnumSource(value = OutputFormat::class)
+    fun `empty path xml`(outputFormat: OutputFormat) {
+        val icon = getResourcePath("xml/ic_without_path.xml")
+        val parserOutput = IconParser.toVector(icon)
+        val output = ImageVectorGenerator.convert(
+            vector = parserOutput.vector,
+            kotlinName = parserOutput.kotlinName,
+            config = ImageVectorGeneratorConfig(
+                packageName = "io.github.composegears.valkyrie.icons",
+                packName = "ValkyrieIcons",
+                nestedPackName = "",
+                outputFormat = outputFormat,
+                generatePreview = false,
+            ),
+        ).content
+
+        val expected = outputFormat.toResourceText(
+            pathToBackingProperty = "kt/backing/WithoutPath.pack.kt",
+            pathToLazyProperty = "kt/lazy/WithoutPath.pack.kt",
+        )
+        assertThat(output).isEqualTo(expected)
+    }
+
+    @ParameterizedTest
+    @EnumSource(value = OutputFormat::class)
+    fun `icon only with path`(outputFormat: OutputFormat) {
+        val icon = getResourcePath("xml/ic_only_path.xml")
+        val parserOutput = IconParser.toVector(icon)
+        val output = ImageVectorGenerator.convert(
+            vector = parserOutput.vector,
+            kotlinName = parserOutput.kotlinName,
+            config = ImageVectorGeneratorConfig(
+                packageName = "io.github.composegears.valkyrie.icons",
+                packName = "ValkyrieIcons",
+                nestedPackName = "",
+                outputFormat = outputFormat,
+                generatePreview = false,
+            ),
+        ).content
+
+        val expected = outputFormat.toResourceText(
+            pathToBackingProperty = "kt/backing/OnlyPath.kt",
+            pathToLazyProperty = "kt/lazy/OnlyPath.kt",
+        )
+        assertThat(output).isEqualTo(expected)
+    }
+
+    @ParameterizedTest
+    @EnumSource(value = OutputFormat::class)
+    fun `icon with path and solid color`(outputFormat: OutputFormat) {
+        val icon = getResourcePath("xml/ic_fill_color_stroke.xml")
+        val parserOutput = IconParser.toVector(icon)
+        val output = ImageVectorGenerator.convert(
+            vector = parserOutput.vector,
+            kotlinName = parserOutput.kotlinName,
+            config = ImageVectorGeneratorConfig(
+                packageName = "io.github.composegears.valkyrie.icons",
+                packName = "ValkyrieIcons",
+                nestedPackName = "",
+                outputFormat = outputFormat,
+                generatePreview = false,
+            ),
+        ).content
+
+        val expected = outputFormat.toResourceText(
+            pathToBackingProperty = "kt/backing/FillColorStroke.kt",
+            pathToLazyProperty = "kt/lazy/FillColorStroke.kt",
+        )
+        assertThat(output).isEqualTo(expected)
+    }
+
+    @ParameterizedTest
+    @EnumSource(value = OutputFormat::class)
+    fun `icon with all path params`(outputFormat: OutputFormat) {
+        val icon = getResourcePath("xml/ic_all_path_params.xml")
+        val parserOutput = IconParser.toVector(icon)
+        val output = ImageVectorGenerator.convert(
+            vector = parserOutput.vector,
+            kotlinName = parserOutput.kotlinName,
+            config = ImageVectorGeneratorConfig(
+                packageName = "io.github.composegears.valkyrie.icons",
+                packName = "ValkyrieIcons",
+                nestedPackName = "",
+                outputFormat = outputFormat,
+                generatePreview = false,
+            ),
+        ).content
+
+        val expected = outputFormat.toResourceText(
+            pathToBackingProperty = "kt/backing/AllPathParams.kt",
+            pathToLazyProperty = "kt/lazy/AllPathParams.kt",
+        )
+        assertThat(output).isEqualTo(expected)
+    }
+
+    @ParameterizedTest
+    @EnumSource(value = OutputFormat::class)
+    fun `icon with several path`(outputFormat: OutputFormat) {
+        val icon = getResourcePath("xml/ic_several_path.xml")
+        val parserOutput = IconParser.toVector(icon)
+        val output = ImageVectorGenerator.convert(
+            vector = parserOutput.vector,
+            kotlinName = parserOutput.kotlinName,
+            config = ImageVectorGeneratorConfig(
+                packageName = "io.github.composegears.valkyrie.icons",
+                packName = "ValkyrieIcons",
+                nestedPackName = "",
+                outputFormat = outputFormat,
+                generatePreview = false,
+            ),
+        ).content
+
+        val expected = outputFormat.toResourceText(
+            pathToBackingProperty = "kt/backing/SeveralPath.kt",
+            pathToLazyProperty = "kt/lazy/SeveralPath.kt",
+        )
+        assertThat(output).isEqualTo(expected)
+    }
+
+    @ParameterizedTest
+    @EnumSource(value = OutputFormat::class)
+    fun `icon with transparent fill color`(outputFormat: OutputFormat) {
+        val icon = getResourcePath("xml/ic_transparent_fill_color.xml")
+        val parserOutput = IconParser.toVector(icon)
+        val output = ImageVectorGenerator.convert(
+            vector = parserOutput.vector,
+            kotlinName = parserOutput.kotlinName,
+            config = ImageVectorGeneratorConfig(
+                packageName = "io.github.composegears.valkyrie.icons",
+                packName = "ValkyrieIcons",
+                nestedPackName = "",
+                outputFormat = outputFormat,
+                generatePreview = false,
+            ),
+        ).content
+
+        val expected = outputFormat.toResourceText(
+            pathToBackingProperty = "kt/backing/TransparentFillColor.kt",
+            pathToLazyProperty = "kt/lazy/TransparentFillColor.kt",
+        )
+        assertThat(output).isEqualTo(expected)
+    }
+}
