+++ components/generator/imagevector/src/test/resources/kt/lazy/Colored.WithoutPath.kt
@@ -1,0 +1,16 @@
+package io.github.composegears.valkyrie.icons.colored
+
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.unit.dp
+import io.github.composegears.valkyrie.icons.ValkyrieIcons
+import kotlin.LazyThreadSafetyMode
+
+val ValkyrieIcons.Colored.WithoutPath: ImageVector by lazy(LazyThreadSafetyMode.NONE) {
+    ImageVector.Builder(
+        name = "Colored.WithoutPath",
+        defaultWidth = 24.dp,
+        defaultHeight = 24.dp,
+        viewportWidth = 18f,
+        viewportHeight = 18f
+    ).build()
+}
