+++ components/generator/imagevector/src/test/resources/kt/lazy/WithoutPath.kt
@@ -1,0 +1,15 @@
+package io.github.composegears.valkyrie.icons
+
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.unit.dp
+import kotlin.LazyThreadSafetyMode
+
+val WithoutPath: ImageVector by lazy(LazyThreadSafetyMode.NONE) {
+    ImageVector.Builder(
+        name = "WithoutPath",
+        defaultWidth = 24.dp,
+        defaultHeight = 24.dp,
+        viewportWidth = 18f,
+        viewportHeight = 18f
+    ).build()
+}
