+++ components/generator/imagevector/src/test/resources/xml/ic_all_path_params.xml
