+++ components/generator/imagevector/src/test/resources/xml/ic_fill_color_stroke.xml
