+++ components/generator/imagevector/src/test/resources/xml/ic_only_path.xml
