+++ components/generator/imagevector/src/test/resources/xml/ic_several_path.xml
