+++ components/generator/imagevector/src/test/resources/xml/ic_transparent_fill_color.xml
