+++ components/generator/imagevector/src/test/resources/xml/ic_without_path.xml
